General refactor of the dialog
