--- lib/overlapping_dialog_route.dart
@@ -1,192 +1,0 @@
-import 'package:flutter/material.dart';
-
-class OverlappingDialogRoute<T> extends RawDialogRoute<T> {
-  OverlappingDialogRoute({
-    required BuildContext context,
-    required WidgetBuilder builder,
-    CapturedThemes? themes,
-    Color? barrierColor = Colors.black54,
-    bool barrierDismissible = true,
-    String? barrierLabel,
-    bool useSafeArea = true,
-    RouteSettings? settings,
-    Offset? anchorPoint,
-  })  : assert(barrierDismissible != null),
-        super(
-          pageBuilder: (BuildContext buildContext, Animation<double> animation,
-              Animation<double> secondaryAnimation) {
-            final Widget pageChild = Builder(builder: builder);
-            Widget dialog = themes?.wrap(pageChild) ?? pageChild;
-            if (useSafeArea) {
-              dialog = SafeArea(child: dialog);
-            }
-            return dialog;
-          },
-          barrierDismissible: barrierDismissible,
-          barrierColor: barrierColor,
-          barrierLabel: barrierLabel ??
-              MaterialLocalizations.of(context).modalBarrierDismissLabel,
-          transitionDuration: const Duration(milliseconds: 150),
-          transitionBuilder: _buildMaterialDialogTransitions,
-          settings: settings,
-          anchorPoint: anchorPoint,
-        ) {
-    localPageBuilder = (BuildContext buildContext, Animation<double> animation,
-        Animation<double> secondaryAnimation) {
-      final Widget pageChild = Builder(builder: builder);
-      Widget dialog = themes?.wrap(pageChild) ?? pageChild;
-      if (useSafeArea) {
-        dialog = SafeArea(child: dialog);
-      }
-      return dialog;
-    };
-  }
-
-  late final RoutePageBuilder localPageBuilder;
-
-  @override
-  Widget buildPage(BuildContext context, Animation<double> animation,
-      Animation<double> secondaryAnimation) {
-    return Semantics(
-      scopesRoute: true,
-      explicitChildNodes: true,
-      child: localPageBuilder(context, animation, secondaryAnimation),
-    );
-  }
-}
-
-class MaterialDialogRoute<T> extends PageRoute<T> {
-  MaterialDialogRoute({
-    required this.pageBuilder,
-    bool barrierDismissible = true,
-    Color? barrierColor = const Color(0x80000000),
-    String? barrierLabel,
-    Duration transitionDuration = const Duration(milliseconds: 200),
-    RouteTransitionsBuilder? transitionBuilder,
-    RouteSettings? settings,
-  })  : assert(barrierDismissible != null),
-        _barrierDismissible = barrierDismissible,
-        _barrierLabel = barrierLabel,
-        _barrierColor = barrierColor,
-        _transitionDuration = transitionDuration,
-        super(settings: settings);
-
-  final RoutePageBuilder pageBuilder;
-
-  @override
-  bool get barrierDismissible => _barrierDismissible;
-  final bool _barrierDismissible;
-
-  @override
-  String? get barrierLabel => _barrierLabel;
-  final String? _barrierLabel;
-
-  @override
-  Color? get barrierColor => _barrierColor;
-  final Color? _barrierColor;
-
-  @override
-  Duration get transitionDuration => _transitionDuration;
-  final Duration _transitionDuration;
-
-  @override
-  Widget buildPage(BuildContext context, Animation<double> animation,
-      Animation<double> secondaryAnimation) {
-    return Semantics(
-      scopesRoute: true,
-      explicitChildNodes: true,
-      child: pageBuilder(context, animation, secondaryAnimation),
-    );
-  }
-
-  @override
-  Widget buildTransitions(BuildContext context, Animation<double> animation,
-      Animation<double> secondaryAnimation, Widget child) {
-    final PageTransitionsTheme theme = Theme.of(context).pageTransitionsTheme;
-    return theme.buildTransitions<T>(
-        this, context, animation, secondaryAnimation, child);
-  }
-
-  @override
-  bool get maintainState => false;
-}
-
-Widget _buildMaterialDialogTransitions(
-    BuildContext context,
-    Animation<double> animation,
-    Animation<double> secondaryAnimation,
-    Widget child) {
-  return FadeTransition(
-    opacity: CurvedAnimation(
-      parent: animation,
-      curve: Curves.easeOut,
-    ),
-    child: child,
-  );
-}
-
-Future<T?> showOverlappingDialog<T>({
-  required BuildContext context,
-  required WidgetBuilder builder,
-  bool barrierDismissible = true,
-  Color? barrierColor = Colors.black54,
-  String? barrierLabel,
-  bool useSafeArea = true,
-  bool useRootNavigator = true,
-  RouteSettings? routeSettings,
-  Offset? anchorPoint,
-}) {
-  assert(builder != null);
-  assert(barrierDismissible != null);
-  assert(useSafeArea != null);
-  assert(useRootNavigator != null);
-  assert(debugCheckHasMaterialLocalizations(context));
-
-  final CapturedThemes themes = InheritedTheme.capture(
-    from: context,
-    to: Navigator.of(
-      context,
-      rootNavigator: useRootNavigator,
-    ).context,
-  );
-
-  return Navigator.of(context, rootNavigator: useRootNavigator)
-      .push<T>(OverlappingDialogRoute<T>(
-    context: context,
-    builder: builder,
-    barrierColor: barrierColor,
-    barrierDismissible: barrierDismissible,
-    barrierLabel: barrierLabel,
-    useSafeArea: useSafeArea,
-    settings: routeSettings,
-    themes: themes,
-    anchorPoint: anchorPoint,
-  ));
-}
-
-Future<T?> showMaterialDialog<T>({
-  required BuildContext context,
-  required WidgetBuilder builder,
-  bool barrierDismissible = true,
-  Color? barrierColor = Colors.black54,
-  String? barrierLabel,
-  bool useRootNavigator = true,
-  RouteSettings? routeSettings,
-}) {
-  assert(builder != null);
-  assert(barrierDismissible != null);
-  assert(useRootNavigator != null);
-  assert(debugCheckHasMaterialLocalizations(context));
-
-  return Navigator.of(context, rootNavigator: useRootNavigator).push<T>(
-    MaterialPageRoute<T>(
-      builder: builder,
-      fullscreenDialog: true,
-      // pageBuilder: (context, animation, secondaryAnimation) => builder(context),
-      // barrierColor: barrierColor,
-      // barrierDismissible: barrierDismissible,
-      // barrierLabel: barrierLabel,
-      settings: routeSettings,
-    ),
-  );
-}
