+++ lib/src/misc/material_dialog_route.dart
@@ -1,0 +1,86 @@
+import 'package:flutter/material.dart';
+
+class MaterialDialogRoute<T> extends PageRoute<T> {
+  MaterialDialogRoute({
+    required this.pageBuilder,
+    bool barrierDismissible = true,
+    Color? barrierColor = const Color(0x80000000),
+    String? barrierLabel,
+    Duration transitionDuration = const Duration(milliseconds: 200),
+    RouteTransitionsBuilder? transitionBuilder,
+    RouteSettings? settings,
+  })  : assert(barrierDismissible != null),
+        _barrierDismissible = barrierDismissible,
+        _barrierLabel = barrierLabel,
+        _barrierColor = barrierColor,
+        _transitionDuration = transitionDuration,
+        super(settings: settings);
+
+  final RoutePageBuilder pageBuilder;
+
+  @override
+  bool get opaque => false;
+
+  @override
+  bool get barrierDismissible => _barrierDismissible;
+  final bool _barrierDismissible;
+
+  @override
+  String? get barrierLabel => _barrierLabel;
+  final String? _barrierLabel;
+
+  @override
+  Color? get barrierColor => _barrierColor;
+  final Color? _barrierColor;
+
+  @override
+  Duration get transitionDuration => _transitionDuration;
+  final Duration _transitionDuration;
+
+  @override
+  Widget buildPage(BuildContext context, Animation<double> animation,
+      Animation<double> secondaryAnimation) {
+    return Semantics(
+      scopesRoute: true,
+      explicitChildNodes: true,
+      child: pageBuilder(context, animation, secondaryAnimation),
+    );
+  }
+
+  @override
+  Widget buildTransitions(BuildContext context, Animation<double> animation,
+      Animation<double> secondaryAnimation, Widget child) {
+    return child;
+    final PageTransitionsTheme theme = Theme.of(context).pageTransitionsTheme;
+    return theme.buildTransitions<T>(
+        this, context, animation, secondaryAnimation, child);
+  }
+
+  @override
+  bool get maintainState => false;
+}
+
+Future<T?> showMaterialDialog<T>({
+  required BuildContext context,
+  required WidgetBuilder builder,
+  bool barrierDismissible = true,
+  Color? barrierColor = Colors.black54,
+  String? barrierLabel,
+  bool useRootNavigator = true,
+  RouteSettings? routeSettings,
+}) {
+  assert(builder != null);
+  assert(barrierDismissible != null);
+  assert(useRootNavigator != null);
+  assert(debugCheckHasMaterialLocalizations(context));
+
+  return Navigator.of(context, rootNavigator: useRootNavigator).push<T>(
+    MaterialDialogRoute<T>(
+      pageBuilder: (context, animation, secondaryAnimation) => builder(context),
+      // barrierColor: barrierColor,
+      barrierDismissible: barrierDismissible,
+      barrierLabel: barrierLabel,
+      settings: routeSettings,
+    ),
+  );
+}
