+++ lib/src/ui/selection_dialog.dart
@@ -1,5 +1,5 @@
-import 'package:country_code_picker/country_code.dart';
-import 'package:country_code_picker/country_localizations.dart';
+import 'package:country_code_picker/src/misc/country_localizations.dart';
+import 'package:country_code_picker/src/model/country_code.dart';
 import 'package:flutter/material.dart';
 
 /// selection dialog used for selection of the country code
